Major refactoring of cdepl
cdepl now offers an interactive shell which allows you to allocate
a session and keep it open. You can re-deploy your application to
that session as many times as you want (or until the session
expires, e.g. if the cluster is using a job based system).
Furthermore, this also provides the foundation for deployment with
deamons like consul.